Fixed warning in Renderer2D

+++ Gaze/Source/Renderer/Renderer2D.cpp
@@ -120,7 +120,8 @@ namespace Gaze {
     void Renderer2D::EndScene() {
         GZ_PROFILE_FUNCTION();
 
-        uint32_t dataSize = (uint8_t *) s_Data.QuadVertexBufferPtr - (uint8_t *) s_Data.QuadVertexBufferBase;
+        uint32_t dataSize = (uint32_t) ((uint8_t *) s_Data.QuadVertexBufferPtr -
+                                        (uint8_t *) s_Data.QuadVertexBufferBase);
         s_Data.QuadVertexBuffer->SetData(s_Data.QuadVertexBufferBase, dataSize);
 
         Flush();
